Fixed some bugs in the interface between Ctrax and the new version of wx, numpy, scipy.
Modified batch processing to allow input of AVIs, etc.

+++ trunk/kcluster2d/kcluster2d_cython.pyx
@@ -449,7 +449,7 @@ def fixsmallpriors(num.ndarray[DTYPE_t,ndim=2] x,
     #    print initcovars[:,:,i]
     #    print 'initcovars[:,:,%d].shape: '%i + str(initcovars[:,:,i].shape)
         
-    cdef num.ndarray[int,ndim=1] smalli
+    cdef num.ndarray[long,ndim=1] smalli
     smalli, = num.where(num.isnan(priors) | (priors < .01))
 
     if DEBUG:
